streams api improvement; minor changes

+++ aui.core/src/AUI/IO/AFileInputStream.h
@@ -55,5 +55,5 @@ public:
 	void seek(std::streamoff offset, Seek dir);
 	void seek(std::streampos pos);
 
-	int read(char* dst, int size) override;
+	size_t read(char* dst, size_t size) override;
 };
