+++ aui.core/src/AUI/Util/ATokenizer.cpp
@@ -327,10 +327,10 @@ glm::vec2 ATokenizer::readVec2()
 	return result;
 }
 
-AString ATokenizer::readString(int n) {
+AString ATokenizer::readString(size_t n) {
     AString result;
     result.reserve(n);
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         result += readChar();
     }
     return result;
